rename Rooms folder and delete unused react-axios from package.json
