+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 
-import Rooms from "./components/Layout/rooms/Rooms.js";
+import Rooms from "./components/Layout/Rooms/Rooms.js";
 import SerachEvent from './components/Common/SerachEvent/SearchEvent.js'
 import Header from './components/Common/Header/Header'
 import Events from './components/Layout/Events/Events'
