+++ package.json
@@ -28,7 +28,6 @@
     "express": "^4.17.1",
     "nodemon": "^2.0.3",
     "pg": "^8.0.2",
-    "react-axios": "^2.0.3",
     "react-router-dom": "^5.1.2"
   },
   "devDependencies": {
